requirements.txt
altgraph==0.17.3
DAWG-Python==0.7.2
docopt==0.6.2
future==0.18.2
importlib-metadata==5.0.0
numpy==1.21.6
pefile==2022.5.30
Pillow==9.0.1
pip==22.2.2
pygame==2.1.2
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyQt5==5.14.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pywin32-ctypes==0.2.0
setuptools==60.9.3
typing_extensions==4.4.0
zipp==3.10.0
